Liza tasklist redirect to taskdetails v1. Successful

+++ Final-Project-Restart-Your-Love/src/main/java/lv/restart/your/love/Final/Project/Restart/Your/Love/service/TaskService.java
@@ -18,7 +18,8 @@ public class TaskService {
     }
 
 
-    public Task findById(int id) {
+    public Task findById(long id) {
+
         return taskRepository.findById(id);
     }
